Move "SelectionFAB" to "collection" dir
