+++ src/components/collection/SelectionFAB.tsx
@@ -11,7 +11,7 @@ import { Fab, Box, styled } from '@mui/material';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import PopupState, { bindMenu, bindTrigger } from 'material-ui-popup-state';
-import { DEFAULT_FAB_STYLE } from '@/components/util/StyledFab';
+import { DEFAULT_FAB_STYLE } from '@/components/util/StyledFab.tsx';
 import { TranslationKey } from '@/typings.ts';
 import { Menu } from '@/components/menu/Menu.tsx';
